% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder.R
\name{DecodeVarint32}
\alias{DecodeVarint32}
\title{Use this decoder version for values which must be limited to 32 bits.}
\usage{
DecodeVarint32(buffer, pos)
}
\description{
Use this decoder version for values which must be limited to 32 bits.
}
\keyword{internal}
