% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_protobuf_messages}
\alias{import_protobuf_messages}
\title{Import Protocol Buffers Descriptors of bayesmix}
\usage{
import_protobuf_messages()
}
\description{
This utility loads in the workspace the protocol buffers descriptors defined
in the \code{bayesmix} library, via \code{RProtoBuf} package. These
descriptors can be used to handle the MCMC chain output of
\code{\link{run_mcmc}} function.
}
