% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{maybe_print_proto_to_file}
\alias{maybe_print_proto_to_file}
\title{Print a protobuf message to file only if input is not a file}
\usage{
maybe_print_proto_to_file(maybe_proto, proto_name = NULL, out_dir = NULL)
}
\description{
If \code{maybe_proto} is a file, returns the file name. If \code{maybe_proto}
is a string representing a message, prints the message to a file and returns
the file name.
}
\keyword{internal}
