% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder.R
\name{VarintDecoder}
\alias{VarintDecoder}
\title{Return a decoder for a basic varint value (does not include tag).}
\usage{
VarintDecoder(mask, result_type)
}
\description{
Decoded values will be bitwise-anded with the given mask before being
returned, e.g. to limit them to 32 bits.  The returned decoder does not take
the usual "end" parameter -- the caller is expected to do bounds checking
after the fact (often the caller can defer such checking until later). The
decoder returns a (value, new_pos) pair.
}
\keyword{internal}
