% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bayesmix.R
\name{build_bayesmix}
\alias{build_bayesmix}
\title{Builds the BayesMix executable}
\usage{
build_bayesmix(
  nproc = ceiling(parallel::detectCores()/2),
  build_dirname = "build"
)
}
\arguments{
\item{nproc}{Number of processes to use for parallel compilation. Thanks to \code{parallel} package,
this parameter defaults to half of the available processes (through \code{\link[parallel]{detectCores}} function)}

\item{build_dirname}{Name for the build directory of BayesMix. Default is "build".}
}
\value{
\code{TRUE} if build is successfull, \code{FALSE} otherwise
}
\description{
After the build, if no error has occurred, it saves the path into the \code{BAYESMIX_EXE} environment variable.
Such variable is defined only when this package is loaded in the R session.
}
